requirements.txt
CherryPy==18.5.0
Pillow>=6.0.0
python-bidi==0.4.2
python-dateutil==2.8.1
requests==2.23.0
Shapely==1.7.0